fix home slider not appear sometimes
